hw5: add saving word in history by opening word
