+++ app/src/main/java/com/example/lightdictionary/domain/SavingWordsRepo.kt
@@ -1,7 +1,7 @@
 package com.example.lightdictionary.domain
 
-import com.example.lightdictionary.data.WordEntity
+import com.example.lightdictionary.data.HistoryEntity
 
 interface SavingWordsRepo {
-    suspend fun saveWord(word: WordEntity)
+    suspend fun saveWord(word: HistoryEntity)
 }
