--- app/src/main/java/com/example/lightdictionary/presenter/MainInteractor.kt
@@ -1,23 +1,0 @@
-package com.example.lightdictionary.presenter
-
-import com.example.lightdictionary.data.LoadWordsState
-import com.example.lightdictionary.domain.LoadingWordsRepo
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.collect
-
-class MainInteractor(private val wordRepo: LoadingWordsRepo) : MainController.Interactor {
-    private val loadWordsStateFlow = MutableStateFlow<LoadWordsState>(LoadWordsState.Loading)
-
-    override suspend fun getData(src: String): StateFlow<LoadWordsState> {
-        return try {
-            wordRepo.getWord(src).collect { words ->
-                loadWordsStateFlow.value = LoadWordsState.Success(words)
-            }
-            loadWordsStateFlow
-        } catch (thr: Throwable) {
-            loadWordsStateFlow.value = LoadWordsState.Error(thr)
-            loadWordsStateFlow
-        }
-    }
-}
